+++ src/protocol/index.ts
@@ -1,1 +1,2 @@
 export { AddressComponentRangeError, MEMORY_REGIONS, MultiSlot, PresetSlot } from "./address";
+export * from "./cc";